Improve interface, randomly display country choices
Major improvements to the system as a whole:
1) Interface gives feedback on whether ( ... listening) or not ( ...)
the system is listening to user speech.
2) There is a button (Listen to me) so that the speech recogniser
is only turned on once the user explicitly clicks on the button.
3) User speech is transcribed in real-time in chrome (ie. interim
results are enabled). Note: Firefox does not show results in real time,
but only once the utterance has completed.
4) There is a designated area, outlined with a dotted line and curved
borders, where all recognition results appear. While the system is
idle/not listening we see a pulsing: (...). Once the system is listening,
we see a pulsing: ( ... listening).
5) The text spoken by the speech synthesiser is displayed in its own
area and it begins with a: ">>".
6) The user can also speak anyhing they want, and these results will also
be displayed in the recognition area (easter eggs disabled for now).
7) The system queries the available speech synthesiser voices on the
users' system, and displays these (countries) as options to choose from.
8) After the user has chosen a country and listened to the sentences,
the system will shuffle and show a random set of countries to choose form.

To implement?
1) a shuffle button (to display a new set of random countries)
2) text input field to input your own country
3) do not display countries double per randomisation (bc its impossible
to choose the second one anyways, plus it looks like a bug)
4) always display the country which appears as the users' default
(because this might be their home country) - note that you can always
speak whatever country (even one that isnt visible on the list)
5) create an array or object with various options for filling in the
adjectives in a random refugee situation. Currently its always showing
the same words.